organizes components folder
