+++ src/components/Catalogue/index.js
@@ -1,0 +1,2 @@
+export { default as ProductItem } from './ProductItem';
+export { default as ProductList } from './ProductList';
