+++ src/components/ShoppingCart/index.js
@@ -1,0 +1,5 @@
+export { default as OrderItem } from './OrderItem';
+export { default as OrderItemList } from './OrderItemList';
+export { default as Product } from './Product';
+export { default as ShoppingCart } from './ShoppingCart';
+export { default as Summary } from './Summary';
